Added typescript to the project
