+++ template/App.tsx
@@ -7,26 +7,25 @@ import {PersistGate} from "redux-persist/integration/react";
 
 import useLocalize from "./src/hooks/useLocalize";
 import Navigation from "./src/navigation/Navigation";
-import store from "./src/redux/store";
+import store from "./src/redux/store"; 
 
 import "react-native-gesture-handler";
 
 const persistor = persistStore(store);
 
 const App = () => {
-    const {setI18nConfig} = useLocalize();
+	const {setI18nConfig} = useLocalize();
+	useEffect(() => {
+		setI18nConfig();
+	}, [setI18nConfig]);
 
-    useEffect(() => {
-        setI18nConfig();
-    }, [setI18nConfig]);
-
-    return (
-        <Provider store={store}>
-            <PersistGate persistor={persistor}>
-                <Navigation />
-            </PersistGate>
-        </Provider>
-    );
+	return (
+		<Provider store={store}>
+			<PersistGate persistor={persistor}>
+				<Navigation />
+			</PersistGate>
+		</Provider>
+	);
 };
 
 export default App;
