+++ template/index.js
@@ -5,6 +5,6 @@
 import {AppRegistry} from "react-native";
 
 import App from "./App";
-import {name as appName} from "./app.json";
+import {name} from "./app.json";
 
-AppRegistry.registerComponent(appName, () => App);
+AppRegistry.registerComponent(name, () => App);
