+++ template/ios/ProjectName.xcworkspace/contents.xcworkspacedata
@@ -7,4 +7,7 @@
    <FileRef
       location = "group:Pods/Pods.xcodeproj">
    </FileRef>
+   <FileRef
+      location = "group:ProjectName.xcodeproj">
+   </FileRef>
 </Workspace>
