+++ template/package.json
@@ -17,7 +17,6 @@
     },
     "dependencies": {
         "@babel/plugin-proposal-export-namespace-from": "^7.18.9",
-        "@digital-art-dealers/react-native-component-lib": "^2.0.0",
         "@react-native-async-storage/async-storage": "^1.18.2",
         "@react-navigation/bottom-tabs": "^6.5.8",
         "@react-navigation/drawer": "^6.6.3",
@@ -53,14 +52,20 @@
         "@react-native/eslint-config": "^0.72.2",
         "@react-native/metro-config": "^0.72.6",
         "@tsconfig/react-native": "^3.0.0",
+        "@types/i18n-js": "^3.8.4",
+        "@types/lodash.memoize": "^4.1.7",
         "@types/metro-config": "^0.76.3",
         "@types/react": "^18.0.24",
         "@types/react-test-renderer": "^18.0.0",
+        "@typescript-eslint/eslint-plugin": "^6.1.0",
+        "@typescript-eslint/parser": "^6.1.0",
         "babel-jest": "^29.2.1",
         "babel-plugin-module-resolver": "^4.1.0",
-        "eslint": "^8.43.0",
+        "eslint": "^8.45.0",
         "eslint-config-airbnb": "^19.0.4",
+        "eslint-plugin-detox": "^1.0.0",
         "eslint-plugin-import": "^2.27.5",
+        "eslint-plugin-jest": "^27.2.3",
         "eslint-plugin-jsx-a11y": "^6.7.1",
         "eslint-plugin-react": "^7.32.2",
         "eslint-plugin-react-hooks": "^4.6.0",
@@ -70,7 +75,7 @@
         "prettier": "^2.8.8",
         "react-test-renderer": "18.2.0",
         "tailwindcss": "^3.3.2",
-        "typescript": "4.8.4"
+        "typescript": "^5.1.6"
     },
     "engines": {
         "node": ">=16"
