+++ template/react-native-vector-icons.d.ts
@@ -1,0 +1,3 @@
+// needed to satisfy typescript issue 
+// Could not find a declaration file for module 'react-native-vector-icons/MaterialIcons'. 
+declare module "react-native-vector-icons/MaterialIcons";
