+++ template/src/hooks/useLocalize/translations.ts
@@ -1,0 +1,12 @@
+import { TranslationType } from "./types";
+
+export const translations: TranslationType = {
+	en: {
+		changeButton: "Change this button",
+		main: "Main",
+	},
+	es: {
+		changeButton: "Cambiar este botón",
+		main: "Principal",
+	},
+};
