+++ template/src/hooks/useLocalize/types.ts
@@ -1,0 +1,7 @@
+export type TranslationKeyType = {
+	[key: string]: string;
+};
+
+export type TranslationType = {
+  [languageTag: string]: TranslationKeyType;
+};
