+++ template/src/navigation/Navigation.tsx
@@ -1,0 +1,34 @@
+import React from "react";
+
+import {NavigationContainer} from "@react-navigation/native";
+import {createStackNavigator} from "@react-navigation/stack";
+
+import LandingScreen from "@screens/LandingScreen";
+import {NAVIGATOR_BOTTOM_TAB, NAVIGATOR_LANDING} from "@utils/screens";
+
+import BottomTabNavigation from "./BottomTabNavigation";
+
+const Stack = createStackNavigator();
+
+const NAVIGATION_OPTIONS = {
+	headerShown: false,
+};
+
+const Navigation = () => (
+	<NavigationContainer>
+		<Stack.Navigator>
+			<Stack.Screen
+				name={NAVIGATOR_LANDING}
+				component={LandingScreen}
+				options={NAVIGATION_OPTIONS}
+			/>
+			<Stack.Screen
+				name={NAVIGATOR_BOTTOM_TAB}
+				component={BottomTabNavigation}
+				options={NAVIGATION_OPTIONS}
+			/>
+		</Stack.Navigator>
+	</NavigationContainer>
+);
+
+export default Navigation;
