+++ template/src/screens/screenTypes.ts
@@ -1,0 +1,6 @@
+import { NAVIGATOR_BOTTOM_TAB, NAVIGATOR_LANDING } from "@utils/screens";
+
+export type NavigatorParamTypes = {
+	[NAVIGATOR_LANDING]: undefined;
+	[NAVIGATOR_BOTTOM_TAB]: undefined;
+};
