--- template/src/utils/screens.js
@@ -1,8 +1,0 @@
-export const NAVIGATORS = {
-    BOTTOM_TAB: "Bottom Tab",
-    LANDING: "Landing Stack",
-};
-
-export const BOTTOM_TAB_SCREENS = {
-    HOME_SCREEN: "Home",
-};
