+++ template/src/utils/screens.ts
@@ -1,0 +1,6 @@
+export const NAVIGATOR_BOTTOM_TAB = "Bottom Tab";
+export const NAVIGATOR_LANDING ="Landing Stack";
+
+export const BOTTOM_TAB_SCREENS = {
+	HOME_SCREEN: "Home",
+};
