--- template/src/utils/translations/translations.js
@@ -1,4 +1,0 @@
-export const TRANSLATIONS = {
-    CHANGE_BUTTON: "changeButton",
-    MAIN: "main",
-};
