+++ template/src/utils/translations/translations.ts
@@ -1,0 +1,6 @@
+import { TranslationKeyType } from "@hooks/useLocalize/types";
+
+export const TRANSLATION_KEYS: TranslationKeyType = {
+	CHANGE_BUTTON: "changeButton",
+	MAIN: "main",
+};
